+++ cart-service/src/main/java/com/project/cart_service/dto/response/CartResponse.java
@@ -1,5 +1,7 @@
 package com.project.cart_service.dto.response;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,5 +13,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class CartResponse {
 
-    private int cartTotal;
+    private BigDecimal cartTotal;
 }
